Moved goa account resources and types to account.go
