+++ design/resources.go
@@ -88,75 +88,6 @@ given work item type can be used in the target of the link.`)
 	})
 })
 
-var _ = a.Resource("user", func() {
-	a.BasePath("/user")
-
-	a.Action("show", func() {
-		a.Security("jwt")
-		a.Routing(
-			a.GET(""),
-		)
-		a.Description("Get the authenticated user")
-		a.Response(d.OK, func() {
-			a.Media(identity)
-		})
-		a.Response(d.BadRequest, JSONAPIErrors)
-		a.Response(d.InternalServerError, JSONAPIErrors)
-		a.Response(d.Unauthorized, JSONAPIErrors)
-	})
-
-})
-
-var _ = a.Resource("identity", func() {
-	a.BasePath("/identities")
-
-	a.Action("list", func() {
-		a.Routing(
-			a.GET(""),
-		)
-		a.Description("List all identities.")
-		a.Response(d.OK, func() {
-			a.Media(identityArray)
-		})
-		a.Response(d.BadRequest, JSONAPIErrors)
-		a.Response(d.InternalServerError, JSONAPIErrors)
-	})
-})
-
-var _ = a.Resource("users", func() {
-	a.BasePath("/users")
-
-	a.Action("show", func() {
-		a.Routing(
-			a.GET("/:id"),
-		)
-		a.Description("Retrieve user for the given ID.")
-		a.Params(func() {
-			a.Param("id", d.String, "id")
-		})
-		a.Response(d.OK, func() {
-			a.Media(identity)
-		})
-		a.Response(d.NotFound, JSONAPIErrors)
-		a.Response(d.InternalServerError, JSONAPIErrors)
-		a.Response(d.BadRequest, JSONAPIErrors)
-
-	})
-
-	a.Action("list", func() {
-		a.Routing(
-			a.GET(""),
-		)
-		a.Description("List all users.")
-		a.Response(d.OK, func() {
-			a.Media(userArray)
-		})
-		a.Response(d.BadRequest, JSONAPIErrors)
-		a.Response(d.InternalServerError, JSONAPIErrors)
-	})
-
-})
-
 var _ = a.Resource("status", func() {
 
 	a.DefaultMedia(ALMStatus)
