+++ migration/sql-files/023-refactor-identities-users.sql
@@ -1,4 +1,4 @@
--- Remove default value of identity IDs. IDs should not be generated by DB. IDs from the keycloak tokens should be used instead
+-- Refactor Identities and Users tables.
 ALTER TABLE identities
     DROP full_name,
     DROP image_url,
