+++ test/login.go
@@ -16,8 +16,8 @@ import (
 func WithIdentity(ctx context.Context, ident account.Identity) context.Context {
 	token := jwt.New(jwt.SigningMethodRS256)
 	token.Claims.(jwt.MapClaims)["uuid"] = ident.ID.String()
-	token.Claims.(jwt.MapClaims)["fullName"] = ident.FullName
-	token.Claims.(jwt.MapClaims)["imageURL"] = ident.ImageURL
+	token.Claims.(jwt.MapClaims)["fullName"] = ident.User.FullName
+	token.Claims.(jwt.MapClaims)["imageURL"] = ident.User.ImageURL
 	return goajwt.WithJWT(ctx, token)
 }
